remove func_service module (for now)
